fix: 로또 개수 int형으로 반환하도록 수정

+++ src/main/java/lotto/Calculator.java
@@ -3,8 +3,8 @@ package lotto;
 import java.util.Map;
 
 public class Calculator {
-    public double getLottoCount(double money) {
-        return money / 1000.0;
+    public int getLottoCount(double money) {
+        return (int)money / 1000;
     }
 
     public Rank calculateRank(int correctCount, boolean bonus) {
